refactor: 挑戦結果のグラフをコンポーネントに抽出
